requirements.txt
apns2==0.7.1
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==3.0.5
django-extensions==2.2.9
django-filter==2.2.0
django-push-notifications==2.0.0
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
drf-yasg==1.17.1
ecdsa==0.15
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.3
phonenumbers==8.12.2
Pillow==7.1.2
psycopg2-binary==2.8.5
py-vapid==1.7.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-jose==3.1.0
pytz==2019.3
pywebpush==1.11.0
PyYAML==5.3.1
qrcode==6.1
requests==2.23.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.5.0
toml==0.10.1
twilio==6.38.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
